Refactor GUI dialogs into separate module
