+++ ui/__init__.py
@@ -1,0 +1,3 @@
+from .api_config import APIConfigDialog
+from .role_config import RoleConfigDialog, ROLE_DESCRIPTIONS
+
